Bit of cleanup with the heartbeat pieces and scene loader code.

+++ Aurora/Framework/AveragingClass.cs
@@ -38,7 +38,7 @@ namespace Aurora.Framework
             avg /= m_list.Count;
             return avg;
 #else
-            return total / count;
+            return total / (float)count;
 #endif
         }
